+++ AVXScan1_multithread_interquery.c
@@ -14,7 +14,7 @@ extern void init_tf(char * data_path, char *query_path);
 int search(int n) {
   int i=0, j=0;
 
-  int base = 0;
+  long base = 0;
   float score;
   __m256i collect_vec, mask;
   __m256 score_vec, t1, t2;
