+++ Scan1_multithread_interquery.c
@@ -12,7 +12,7 @@
 extern void init_tf(char * data_path, char *query_path);
 int search(int n) {
   int i=0, j=0;
-  int base=0;
+  long base=0;
   float score;
   int t;
   heap h;
