+++ Scan2_multithread_intraquery.c
@@ -13,7 +13,7 @@ struct arg_struct {
     int topic;
     int startidx;
     int endidx;
-    int base;
+    long base;
     heap* h;
 };
 
@@ -26,7 +26,7 @@ int search(struct arg_struct *arg) {
   heap_create(h,0,NULL);
 
   int i=0, j=0;
-  int base = arg->base;
+  long base = arg->base;
   float score;
   int t;
   float* min_key;
